Assignment-4 Modifications
