+++ assignment-4/src/components/ConfirmPassword.js
@@ -1,11 +1,9 @@
-import React from 'react';
-
 const ConfirmPassword = (props) => {
     return (
         <div>
             <h4>Confirm Password :</h4>
             <input type = "password" name = "confirmPassword" onChange = {props.onChange} placeholder = "Confirm Password..." />
-            <br /><br />
+            <p style = {{ color: 'red' }} id = "confirmPwdError" />
         </div>
     );
 }
