+++ assignment-4/src/components/Email.js
@@ -1,10 +1,9 @@
-import React from 'react';
-
 const Email = (props) => {
     return (
         <div>
             <h4>Email :</h4>
             <input type = "text" name = "email" onChange = {props.onChange} placeholder = "xyz@gmail.com" />
+            <p style = {{ color: 'red' }} id = "emailError" />
         </div>
     )
 }
