+++ assignment-4/src/components/Password.js
@@ -1,10 +1,9 @@
-import React from 'react';
-
 const Password = (props) => {
     return (
         <div>
             <h4>Password :</h4>
             <input type = "password" name = "password" onChange = {props.onChange} placeholder = "Enter Password..." />
+            <p style = {{ color: 'red' }} id = "pwdError" />
         </div>
     )
 }
