+++ assignment-4/src/components/ProfilePic.js
@@ -1,10 +1,9 @@
-import React from 'react';
-
 const ProfilePic = (props) => {
     return (
         <div>
             <h4>Upload Profile Pic :</h4>
             <input type = "file" name = "profilePicUrl" onChange = {props.onChange} />
+            <p style = {{ color: 'red' }} id = "profilePicError" />
         </div>
     )
 }
